Use in_txn for migration so we get the commit logging
